Refactor configuration of SSL to allow override with custom config
