+++ src/main/java/com/github/dockerjava/api/command/DockerCmdExecFactory.java
@@ -3,8 +3,6 @@ package com.github.dockerjava.api.command;
 import java.io.Closeable;
 import java.io.IOException;
 
-import javax.net.ssl.SSLContext;
-
 import com.github.dockerjava.core.DockerClientConfig;
 import com.github.dockerjava.core.RemoteApiVersion;
 
@@ -116,8 +114,6 @@ public interface DockerCmdExecFactory extends Closeable {
     ConnectToNetworkCmd.Exec createConnectToNetworkCmdExec();
 
     DisconnectFromNetworkCmd.Exec createDisconnectFromNetworkCmdExec();
-
-    DockerCmdExecFactory withSSLContext(SSLContext sslContext);
 
     @Override
     void close() throws IOException;
